+++ modules/apps/foundation/user-associated-data/user-associated-data-web/src/main/java/com/liferay/user/associated/data/web/internal/util/UADApplicationSummaryHelper.java
@@ -172,7 +172,7 @@ public class UADApplicationSummaryHelper {
 		).map(
 			key -> _uadRegistry.getUADEntityAggregator(key)
 		).mapToInt(
-			uadEntityAggregator -> uadEntityAggregator.count(userId)
+			uadEntityAggregator -> (int)uadEntityAggregator.count(userId)
 		).sum();
 	}
